wired new predictor library into console app

+++ PredictorLibrary/FiguresForYear.cs
@@ -1,3 +1,3 @@
 namespace PredictorLibrary;
 
-public record FiguresForYear(int Index, int Year, int Age, decimal AmountAtStartOfYear, decimal AmountAtEndOfYear);
+public record FiguresForYear(int Index, int Year, int Age, double AmountAtStartOfYear, double AmountAtEndOfYear);
